# PhotographyBack/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Mathis Petit",
        "description": "Documentation API pour le projet de photographie",
        "contact": {
            "email": "mathispetitfrpro@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Lister tous les utilisateurs (admin uniquement)",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher les informations d'un utilisateur spécifique",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour les informations d'un utilisateur",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur mis à jour avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/liked-photos": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Récupérer les photos likées par l'utilisateur connecté",
                "operationId": "4d96110def7a58267f808dc0ec143f7a",
                "responses": {
                    "200": {
                        "description": "Photos likées par l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Photo"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "description": "ID of the category",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the category",
                        "type": "string",
                        "example": "Landscape"
                    },
                    "created_at": {
                        "description": "Timestamp when the category was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-04T12:34:56.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the category was updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-05T12:34:56.000000Z"
                    },
                    "photos": {
                        "description": "List of photos belonging to the category",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            },
            "Partner": {
                "title": "Partner",
                "description": "Modèle représentant un partenaire photographe dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique du partenaire",
                        "type": "integer"
                    },
                    "bio": {
                        "description": "Biographie du partenaire",
                        "type": "string"
                    },
                    "website": {
                        "description": "Site web du partenaire",
                        "type": "string"
                    },
                    "social_links": {
                        "description": "Liens vers les réseaux sociaux du partenaire",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur associé au partenaire",
                        "type": "integer"
                    },
                    "photos": {
                        "description": "Liste des photos associées à ce partenaire",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Photo": {
                "title": "Photo",
                "description": "Modèle représentant une photo dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique de la photo",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Titre de la photo",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description de la photo",
                        "type": "string"
                    },
                    "image_path": {
                        "description": "Chemin vers l'image de la photo",
                        "type": "string"
                    },
                    "likes": {
                        "description": "Nombre de likes que la photo a reçu",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur ayant téléchargé la photo",
                        "type": "integer"
                    },
                    "categories": {
                        "description": "Liste des catégories associées à cette photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "likedByUsers": {
                        "description": "Liste des utilisateurs ayant liké la photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Modèle représentant un utilisateur dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique de l'utilisateur",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom de l'utilisateur",
                        "type": "string"
                    },
                    "email": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "description": "Rôle de l'utilisateur (user, partner, admin)",
                        "type": "string"
                    },
                    "profile_photo_url": {
                        "description": "URL de la photo de profil de l'utilisateur",
                        "type": "string"
                    },
                    "photos": {
                        "description": "Liste des photos associées à cet utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    },
                    "partner": {
                        "$ref": "#/components/schemas/Partner"
                    },
                    "likedPhotos": {
                        "description": "Liste des photos likées par l'utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}